kvserver: mode TestDecommission to kv-owned file
`client_raft_test.go` is owned by the Replication team.

Noticed in #106096.

Epic: none
Release note: None
